refactor: App.js components 폴더로 이동
